feature/APCOR-555 to release/APCOR-555

+++ src/main/java/com/ath/adminefectivo/entities/TarifasOperacion.java
@@ -74,7 +74,7 @@ public class TarifasOperacion {
 	private String comisionAplicar;
 
 	@Column(name = "VALOR_TARIFA")
-	private int valorTarifa; 
+	private Double valorTarifa; 
 
 	@Column(name = "ESTADO", nullable = true)
 	private int estado;
